Refactor External Scans to Foreign Scans in Orca
This commit is in preperation for adding foreign scan support in Orca

- Rename ExternalScans to ForeignScans
- Get rid of CMDRelationExternalGPDB. This had information specific to
external scans, but Orca didn't use any of this information. Instead,
we populate this necessary information in the translator. It's possible
we may be able to do some optimizations such as predicate pushdown,
using indexes, etc in the future; however, the interface to the FDW api
is through planner structures so they'd have to be rewritten anyway.
It's more likely that we keep this in the translator.

(cherry picked from commit ecd0c5c788c6b04ac5a08ce4e4907dc705d070ef)