Fix test_consume_xids where it consumes +1 to what's intended
test_consume_xids(n int) is supposed to consume n xids but it
actually consumes n+1. The reason is that we initialize 'xid' to
be the next xid but during advancing it we'll rewrite it to be
the xid that we just allocated (the "current xid"). Fixing it by
making its meaning consistent.

(cherry picked from commit 93d5419ffc9594364abb8c21b84e6169dbefaa47)