Fix relcache lookup in Orca when selecting from sequence
In d0f8614d1f1b18eb9101bb18b8e9192709bd670c, we removed an if statement
that was previously catching and falling back to planner if a table AM
was not recognized. Add this back, as there are cases of an unrecognized
AM and otherwise we'll crash.

(cherry picked from commit 049ffdbf6ac9aaf7fd93f9382dcc87031c30ee23)