ao/co: Ban speculative insert in parse analysis
We don't support speculative insert (INSERT ON CONFLICT) for AO/CO
tables yet.

This commit does two things:

(1) Confirms that its okay to keep the speculative insert APIs
unimplemented, by removing the FIXMEs and adding ereport().

(2) Instead of relying on the ERROR at the table AM layer, it introduces
a check at the parse analysis stage to perform the ban. This is
important as we can't rely on the table AM layer always. For instance,
with DO NOTHING, if there is no conflict, the table AM routine is
invoked. However, if there is a conflict, then the routine doesn't get
invoked and the customer would see:

INSERT INTO spec_insert VALUES(1, 3) ON CONFLICT(i) DO NOTHING;
INSERT 0 0

This would lead to a false impression that it is supported. So, we ban
it at the parse analysis stage.

(cherry picked from commit df7b3b69753ba0d12912d33ae21c87a5f9008104)