Not grab distributed snapshot if it's direct dispatch (#13991)
Currently, we will create a distributed snapshot in the function GetSnapshotData() if
we are QD, and we will iterate procArray again to get the global xmin/xmax/xip.

But if the current query could be dispatched to a single segment directly, which means
it's a direct dispatch, there is no need to create a distributed snapshot, the local snapshot
is enough.

(cherry picked from commit c15410ecbd4606d67668da50eba4b373e2e4dadf)