Remove result tuple bound push down if qual
Fixes #10803

Issue:
https://github.com/greenplum-db/gpdb/issues/10803
When ORCA generates a RESULT node with a LIMIT parent, the query returns a
subset of the actual result.

Root cause:
When ORCA generates a RESULT node with a LIMIT parent, the parent node's tuple
bound is pushed down to the RESULT node's child node. This is expected unless
the RESULT node has a filter. The filter is supposed to be applied before the
tuple bound, instead it's currently applied after the tuple bound. This leads to
the query returning a subset of the actual result.

Example:
Consider this plan shape --

->  Limit
	 ->  Result
	       Filter: (SubPlan 1)
	       ->  Sort
		     ...
	       SubPlan 1
		     ...

The RESULT node has a LIMIT parent and a SORT child. The correct procedure
should be: (1) sort the relation, (2) apply the filter to the sorted relation
and (3) apply limit to the filtered relation. But instead, the limit was applied
before the filter to the relation.

Solution:
When executing the plan, we allow tuple bound push down only if the RESULT node
DOES NOT has a filter condition. If the RESULT node has a non-NULL QUAL, we
don't apply the tuple bound to its child node.

(cherry picked from commit 9c9d8b99f8661d28e2f2e465145d964da37ef027)