fix the issue of cannot create temporary table like existing table with comments (#14742)
This PR is trying to fix the issue of #14649.

When QD generates query tree and serializes it to string and sends it to QE,
and QE will deserialize it to a plan tree. In this process, Greenplum seems not
to consider the difference between NULL and an empty string, so if the original
value is a NULL, QE may deserialize it to an empty string, which could lead error
happened.

It's hard to modify the function of deserializeNode(), since Greenplum does not
have the mechanism to clarify NULL and empty string during serialization and
deserialization.

So this PR is just fixing the special issue, not for general using.

(cherry picked from commit 77cdd310fa84d8af3d7eaadc1b0f61dba9f916ef)