Fix value in the cpu_usage column returned from the pg_resgroup_get_status() function (#14959)
In the case of insert-select, the cpu_usage field contains information about
master only. This leads to an error at the moment of analyzing the table, as the
contents of the field do not correspond to json format.
The error was fixed by setting the proexeclocation property of the function
to i (PROEXECLOCATION_INITPLAN).

https://github.com/greenplum-db/gpdb/issues/14154
(cherry picked from commit fdf47fa2e31c06704d67aba61e518e27855f616b)

Adaptations for greengage/next:
* catversion.h: the commit changes pg_proc.dat, so CATALOG_VERSION_NO
  is set to a fresh porting-date value 302608081 rather than either
  side of the conflict.