Orca FIXME: Improve stats calculation during static partition selection (#14958)
With the new partitioning architecture, we pass through the constraint
expressions on the partitioning key to the dynamic table scan. Therefore, we
don't need to generate the predicate on the partition key, but instead we can
retrieve it from m_partition_cnstrs_disj->PexprScalar(mp). I didn't modify any
of the dynamic elimination logic--this is only for the static elimination case.

The goal of this concept is to properly estimate the cardinality of scans on
partitioned tables. In 6X, we didn't have access to individual constraints on
individual partitions, so we had to estimate and assume uniformity. We
extracted the partition key and assumed that was the "constraint" that we were
scanning on. We now have the actual constraint for the partitions, so we use
those instead.

When costing a table scan, we use the stats of the entire table. Scanning 1
million tuples costs more than 10 tuples, even if we're projecting 10 tuples in
each case. For non-partitioned tables, we simply get the stats of the base
table. For partitioned tables within orca, we want the stats of the partitioned
table after we've done static and dynamic elimination. In the previous
partitioning architecture, we didn't perform static elimination within Orca and
didn't have access to the constraints of individual partitions. If we had a
predicate where pkey=2, we had to estimate how many rows were in the partition
containing pkey=2. Now we have those constraints, and can use these constraints
for more accurate estimates.

There were a couple of cardinality changes with this, but they're more accurate
as far as I can tell. The change in EffectOfLocalPredOnJoin2.mdp is a better
estimate, since the partitions being scanned don't have any statistics
according to the mdp.

(cherry picked from commit 78b7bd8ada85467063b17361216b17c0533f9450)