Resolve merge fixmes from syncrep.c
The implementation questioned is not new and has been there for
GPDB4,5,6 and 7. Longer term goal is to make it better but nothing in
short term to fix here.

Plus, the fixme in SyncRepGetSyncStandbys() can't be done as
standbconfig is not created or populated for
coordinator-standby. Hence, the code needs to live in
SyncRepGetSyncStandbys(). Once proper implementation is done for
coodinator and standby then this part can be tackled as well.

(cherry picked from commit 766a062844e8819fbd7802c6cf91ff330661c666)

Adaptation for next: only the comment above SyncRepWaitForLSN() is carried
over.  The second hunk dropped a GPDB_12_MERGE_FIXME sitting in
SyncRepGetSyncStandbys(), but upstream f332241a60a ("Fix race conditions in
synchronous standby management.", PG13) replaced that whole function family
with SyncRepGetCandidateStandbys(), so next has no such comment to remove; the
IS_QUERY_DISPATCHER special case it referred to now lives in
SyncRepGetCandidateStandbys().